Add user interface and change logic
Removed logic that plays exactly five rounds and made it so the first
player to reach 5 points wins. Added multiple methods to display
information to user interface. They include displayChoices()
getFinalMessage() and displayResults(). Added addButtonListeners to add
event listeners to buttons on user interface and restartGame() as a
callback function to the button that reloads webpage.